Restore previous `ButtonSolid`, add new `ButtonsLegacy` page
