+++ example/src/navigation/navigator.tsx
@@ -19,6 +19,7 @@ import { DSAdvice } from "../pages/Advice";
 import { DSAlert } from "../pages/Alert";
 import { Badges } from "../pages/Badges";
 import { Buttons } from "../pages/Buttons";
+import { ButtonsLegacy } from "../pages/ButtonsLegacy";
 import { Collapsible } from "../pages/Collapsible";
 import { Colors } from "../pages/Colors";
 import { FooterActionsEmptyStateScreen } from "../pages/FooterActionsEmptyStateScreen";
@@ -189,6 +190,14 @@ const AppNavigator = () => {
             component={Buttons}
             options={{
               headerTitle: APP_ROUTES.COMPONENTS.BUTTONS.title,
+              headerBackTitleVisible: false
+            }}
+          />
+          <Stack.Screen
+            name={APP_ROUTES.COMPONENTS.BUTTONS_LEGACY.route}
+            component={ButtonsLegacy}
+            options={{
+              headerTitle: APP_ROUTES.COMPONENTS.BUTTONS_LEGACY.title,
               headerBackTitleVisible: false
             }}
           />
