+++ example/src/navigation/params.ts
@@ -13,6 +13,7 @@ export type AppParamsList = {
   [DESIGN_SYSTEM_ROUTES.COMPONENTS.IMAGE.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.COMPONENTS.OTP_INPUT.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.COMPONENTS.BUTTONS.route]: undefined;
+  [DESIGN_SYSTEM_ROUTES.COMPONENTS.BUTTONS_LEGACY.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.COMPONENTS.BADGE.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.COMPONENTS.SELECTION.route]: undefined;
   [DESIGN_SYSTEM_ROUTES.COMPONENTS.ADVICE.route]: undefined;
