+++ example/src/pages/ButtonsLegacy.tsx
@@ -1,6 +1,4 @@
 import {
-  Button,
-  ButtonColor,
   ButtonLink,
   ButtonOutline,
   ButtonSolid,
@@ -17,7 +15,7 @@ import {
   VStack,
   hexToRgba
 } from "@pagopa/io-app-design-system";
-import React, { Fragment, useState } from "react";
+import React, { useState } from "react";
 import { Alert, StyleSheet, View } from "react-native";
 import { ComponentViewerBox } from "../components/ComponentViewerBox";
 import { Screen } from "../components/Screen";
@@ -41,10 +39,7 @@ const onButtonPress = () => {
   Alert.alert("Alert", "Action triggered");
 };
 
-const buttonColors: Array<ButtonColor> = ["primary", "danger", "contrast"];
-
-// sonarjs/cognitive-complexity
-export const Buttons = () => (
+export const ButtonsLegacy = () => (
   <Screen>
     {/* The title should be dynamic, got from the route object */}
     <H2
@@ -57,21 +52,18 @@ export const Buttons = () => (
     </H2>
     <ComponentViewerBox name="ButtonSolid · Primary Variant (using Pressable API)">
       <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="solid"
+        <ButtonSolid
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button"}
           onPress={onButtonPress}
         />
-        <Button
-          variant="solid"
+        <ButtonSolid
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button"}
           icon="qrCode"
           onPress={onButtonPress}
         />
-        <Button
-          variant="solid"
+        <ButtonSolid
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button"}
           icon="qrCode"
@@ -79,8 +71,7 @@ export const Buttons = () => (
           onPress={onButtonPress}
         />
         <View style={{ alignSelf: "center" }}>
-          <Button
-            variant="solid"
+          <ButtonSolid
             accessibilityHint="Tap to trigger test alert"
             label={"Primary button (centered)"}
             onPress={onButtonPress}
@@ -92,8 +83,7 @@ export const Buttons = () => (
       {/* Let's force `alignItems: "flex-start"`
       to test if `fullWidth` is managed correctly */}
       <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="solid"
+        <ButtonSolid
           fullWidth
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button (full width)"}
@@ -106,15 +96,13 @@ export const Buttons = () => (
     </ComponentViewerBox>
     <ComponentViewerBox name="ButtonSolid · Primary, disabled">
       <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="solid"
+        <ButtonSolid
           disabled
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button (disabled)"}
           onPress={onButtonPress}
         />
-        <Button
-          variant="solid"
+        <ButtonSolid
           disabled
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button (disabled)"}
@@ -123,25 +111,23 @@ export const Buttons = () => (
         />
       </VStack>
     </ComponentViewerBox>
+
     <ComponentViewerBox name="ButtonSolid · Danger variant">
       <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="solid"
+        <ButtonSolid
           color="danger"
           label={"Danger button"}
           onPress={onButtonPress}
           accessibilityHint="Tap to trigger test alert"
         />
-        <Button
-          variant="solid"
+        <ButtonSolid
           color="danger"
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button"}
           icon="trashcan"
           onPress={onButtonPress}
         />
-        <Button
-          variant="solid"
+        <ButtonSolid
           color="danger"
           accessibilityHint="Tap to trigger test alert"
           label={"Primary button"}
@@ -152,8 +138,7 @@ export const Buttons = () => (
       </VStack>
     </ComponentViewerBox>
     <ComponentViewerBox name="ButtonSolid · Danger, full width">
-      <Button
-        variant="solid"
+      <ButtonSolid
         fullWidth
         color="danger"
         accessibilityHint="Tap to trigger test alert"
@@ -161,18 +146,17 @@ export const Buttons = () => (
         onPress={onButtonPress}
       />
     </ComponentViewerBox>
+
     <ComponentViewerBox name="ButtonSolid · Danger, disabled">
       <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="solid"
+        <ButtonSolid
           color="danger"
           disabled
           accessibilityHint="Tap to trigger test alert"
           label={"Danger button (disabled)"}
           onPress={onButtonPress}
         />
-        <Button
-          variant="solid"
+        <ButtonSolid
           color="danger"
           disabled
           accessibilityHint="Tap to trigger test alert"
@@ -182,22 +166,21 @@ export const Buttons = () => (
         />
       </VStack>
     </ComponentViewerBox>
+
     <View style={styles.primaryBlock}>
       <ComponentViewerBox
         name="ButtonSolid · Contrast variant"
         colorMode="dark"
       >
         <VStack space={16} style={{ alignItems: "flex-start" }}>
-          <Button
-            variant="solid"
+          <ButtonSolid
             color="contrast"
             label={"Contrast button"}
             onPress={onButtonPress}
             accessibilityHint="Tap to trigger test alert"
           />
 
-          <Button
-            variant="solid"
+          <ButtonSolid
             color="contrast"
             label={"Contrast button"}
             icon="add"
@@ -205,8 +188,7 @@ export const Buttons = () => (
             accessibilityHint="Tap to trigger test alert"
           />
 
-          <Button
-            variant="solid"
+          <ButtonSolid
             color="contrast"
             label={"Contrast button"}
             icon="add"
@@ -221,8 +203,7 @@ export const Buttons = () => (
         name="ButtonSolid · Contrast, full width"
         colorMode="dark"
       >
-        <Button
-          variant="solid"
+        <ButtonSolid
           fullWidth
           color="contrast"
           label={"Contrast button"}
@@ -235,8 +216,7 @@ export const Buttons = () => (
         name="ButtonSolid · Contrast, full width, loading state"
         colorMode="dark"
       >
-        <Button
-          variant="solid"
+        <ButtonSolid
           fullWidth
           loading
           color="contrast"
@@ -252,8 +232,7 @@ export const Buttons = () => (
         last
       >
         <VStack space={16} style={{ alignItems: "flex-start" }}>
-          <Button
-            variant="solid"
+          <ButtonSolid
             disabled
             color="contrast"
             label={"Contrast button (disabled)"}
@@ -261,234 +240,7 @@ export const Buttons = () => (
             accessibilityHint="Tap to trigger test alert"
           />
 
-          <Button
-            variant="solid"
-            disabled
-            color="contrast"
-            label={"Contrast button (disabled)"}
-            icon="add"
-            onPress={onButtonPress}
-            accessibilityHint="Tap to trigger test alert"
-          />
-        </VStack>
-      </ComponentViewerBox>
-    </View>
-    <H2
-      style={{
-        marginBottom: 16,
-        paddingTop: IOVisualCostants.appMarginDefault
-      }}
-    >
-      Button (Outline variant)
-    </H2>
-    <ComponentViewerBox name="ButtonSolid · Primary Variant (using Pressable API)">
-      <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="outline"
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button"}
-          onPress={onButtonPress}
-        />
-        <Button
-          variant="outline"
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button"}
-          icon="qrCode"
-          onPress={onButtonPress}
-        />
-        <Button
-          variant="outline"
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button"}
-          icon="qrCode"
-          iconPosition="end"
-          onPress={onButtonPress}
-        />
-        <View style={{ alignSelf: "center" }}>
-          <Button
-            variant="outline"
-            accessibilityHint="Tap to trigger test alert"
-            label={"Primary button (centered)"}
-            onPress={onButtonPress}
-          />
-        </View>
-      </VStack>
-    </ComponentViewerBox>
-    <ComponentViewerBox name="ButtonSolid · Primary, Full width">
-      {/* Let's force `alignItems: "flex-start"`
-      to test if `fullWidth` is managed correctly */}
-      <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="outline"
-          fullWidth
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button (full width) with a loong looong label"}
-          onPress={onButtonPress}
-        />
-      </VStack>
-    </ComponentViewerBox>
-    <ComponentViewerBox name="ButtonSolid · Primary · Full width, loading state">
-      <LoadingOutlineButtonExample />
-    </ComponentViewerBox>
-    <ComponentViewerBox name="ButtonSolid · Primary, disabled">
-      <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="outline"
-          disabled
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button (disabled)"}
-          onPress={onButtonPress}
-        />
-        <Button
-          variant="outline"
-          disabled
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button (disabled)"}
-          icon="qrCode"
-          onPress={onButtonPress}
-        />
-      </VStack>
-    </ComponentViewerBox>
-    <ComponentViewerBox name="ButtonSolid · Danger variant">
-      <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="outline"
-          color="danger"
-          label={"Danger button"}
-          onPress={onButtonPress}
-          accessibilityHint="Tap to trigger test alert"
-        />
-        <Button
-          variant="outline"
-          color="danger"
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button"}
-          icon="trashcan"
-          onPress={onButtonPress}
-        />
-        <Button
-          variant="outline"
-          color="danger"
-          accessibilityHint="Tap to trigger test alert"
-          label={"Primary button"}
-          icon="trashcan"
-          iconPosition="end"
-          onPress={onButtonPress}
-        />
-      </VStack>
-    </ComponentViewerBox>
-    <ComponentViewerBox name="ButtonSolid · Danger, full width">
-      <Button
-        variant="outline"
-        fullWidth
-        color="danger"
-        accessibilityHint="Tap to trigger test alert"
-        label={"Danger button (full width)"}
-        onPress={onButtonPress}
-      />
-    </ComponentViewerBox>
-    <ComponentViewerBox name="ButtonSolid · Danger, disabled">
-      <VStack space={16} style={{ alignItems: "flex-start" }}>
-        <Button
-          variant="outline"
-          color="danger"
-          disabled
-          accessibilityHint="Tap to trigger test alert"
-          label={"Danger button (disabled)"}
-          onPress={onButtonPress}
-        />
-        <Button
-          variant="outline"
-          color="danger"
-          disabled
-          accessibilityHint="Tap to trigger test alert"
-          label={"Danger button (disabled)"}
-          icon="trashcan"
-          onPress={onButtonPress}
-        />
-      </VStack>
-    </ComponentViewerBox>
-    <View style={styles.primaryBlock}>
-      <ComponentViewerBox
-        name="ButtonSolid · Contrast variant"
-        colorMode="dark"
-      >
-        <VStack space={16} style={{ alignItems: "flex-start" }}>
-          <Button
-            variant="outline"
-            color="contrast"
-            label={"Contrast button"}
-            onPress={onButtonPress}
-            accessibilityHint="Tap to trigger test alert"
-          />
-
-          <Button
-            variant="outline"
-            color="contrast"
-            label={"Contrast button"}
-            icon="add"
-            onPress={onButtonPress}
-            accessibilityHint="Tap to trigger test alert"
-          />
-
-          <Button
-            variant="outline"
-            color="contrast"
-            label={"Contrast button"}
-            icon="add"
-            iconPosition="end"
-            onPress={onButtonPress}
-            accessibilityHint="Tap to trigger test alert"
-          />
-        </VStack>
-      </ComponentViewerBox>
-
-      <ComponentViewerBox
-        name="ButtonSolid · Contrast, full width"
-        colorMode="dark"
-      >
-        <Button
-          variant="outline"
-          fullWidth
-          color="contrast"
-          label={"Contrast button"}
-          onPress={onButtonPress}
-          accessibilityHint="Tap to trigger test alert"
-        />
-      </ComponentViewerBox>
-
-      <ComponentViewerBox
-        name="ButtonSolid · Contrast, full width, loading state"
-        colorMode="dark"
-      >
-        <Button
-          variant="outline"
-          fullWidth
-          loading
-          color="contrast"
-          label={"Contrast button"}
-          onPress={onButtonPress}
-          accessibilityHint="Tap to trigger test alert"
-        />
-      </ComponentViewerBox>
-
-      <ComponentViewerBox
-        name="ButtonSolid · Contrast, disabled"
-        colorMode="dark"
-        last
-      >
-        <VStack space={16} style={{ alignItems: "flex-start" }}>
-          <Button
-            variant="outline"
-            disabled
-            color="contrast"
-            label={"Contrast button (disabled)"}
-            onPress={onButtonPress}
-            accessibilityHint="Tap to trigger test alert"
-          />
-
-          <Button
-            variant="outline"
+          <ButtonSolid
             disabled
             color="contrast"
             label={"Contrast button (disabled)"}
@@ -499,7 +251,9 @@ export const Buttons = () => (
         </VStack>
       </ComponentViewerBox>
     </View>
+
     <VSpacer size={40} />
+
     <H2 style={{ marginBottom: 16, marginTop: 16 }}>ButtonOutline</H2>
     <ComponentViewerBox name="ButtonOutline · Primary variant (using Pressable API)">
       <VStack space={16} style={{ alignItems: "flex-start" }}>
@@ -560,6 +314,7 @@ export const Buttons = () => (
         />
       </VStack>
     </ComponentViewerBox>
+
     <View style={styles.primaryBlock}>
       <ComponentViewerBox
         name="ButtonOutline · Contrast variant"
@@ -629,115 +384,9 @@ export const Buttons = () => (
         </VStack>
       </ComponentViewerBox>
     </View>
-    <VSpacer size={40} />
-    <H2 style={{ marginBottom: 16, marginTop: 16 }}>Button (link variant)</H2>
-    <VStack space={32}>
-      {buttonColors.map((color, index) => {
-        const isContrast = color === "contrast";
-
-        return (
-          <Fragment key={`${color}-link-variant-${index}`}>
-            <View style={isContrast ? styles.primaryBlock : {}}>
-              <ComponentViewerBox
-                name={`Button · Link variant, ${color} color`}
-                colorMode={isContrast ? "dark" : undefined}
-              >
-                <Button
-                  color={color}
-                  variant="link"
-                  accessibilityHint="Tap to trigger test alert"
-                  label={"Primary button"}
-                  onPress={onButtonPress}
-                />
-
-                <VSpacer size={16} />
-
-                <Button
-                  color={color}
-                  variant="link"
-                  accessibilityHint="Tap to trigger test alert"
-                  label={"Primary button"}
-                  icon="starEmpty"
-                  onPress={onButtonPress}
-                />
-
-                <VSpacer size={16} />
-
-                <Button
-                  color={color}
-                  variant="link"
-                  accessibilityHint="Tap to trigger test alert"
-                  label={"Primary button"}
-                  icon="starEmpty"
-                  iconPosition="end"
-                  onPress={onButtonPress}
-                />
-
-                <VSpacer size={16} />
-
-                <View style={{ alignSelf: "center" }}>
-                  <Button
-                    color={color}
-                    variant="link"
-                    accessibilityHint="Tap to trigger test alert"
-                    label={"Primary button (centered)"}
-                    onPress={onButtonPress}
-                  />
-                </View>
-              </ComponentViewerBox>
-              <ComponentViewerBox
-                name="ButtonLink · Stress test"
-                colorMode={isContrast ? "dark" : undefined}
-              >
-                <View style={{ alignSelf: "center" }}>
-                  <Button
-                    color={color}
-                    variant="link"
-                    textAlign="center"
-                    /* Don't set limits on maximum number of lines */
-                    numberOfLines={0}
-                    accessibilityHint="Tap to trigger test alert"
-                    label={"Primary button (centered) with a very looong text"}
-                    onPress={onButtonPress}
-                  />
-                </View>
-              </ComponentViewerBox>
-              <ComponentViewerBox
-                name={`Button · Link variant, ${color} color, disabled`}
-                colorMode={isContrast ? "dark" : undefined}
-                last
-              >
-                <View>
-                  <Button
-                    color={color}
-                    variant="link"
-                    disabled
-                    accessibilityHint="Tap to trigger test alert"
-                    label={"Primary button (disabled)"}
-                    onPress={onButtonPress}
-                  />
-
-                  <VSpacer size={16} />
-
-                  <Button
-                    color={color}
-                    variant="link"
-                    disabled
-                    accessibilityHint="Tap to trigger test alert"
-                    label={"Primary button (disabled)"}
-                    icon="starEmpty"
-                    iconPosition="end"
-                    onPress={onButtonPress}
-                  />
-                </View>
-              </ComponentViewerBox>
-            </View>
-          </Fragment>
-        );
-      })}
-    </VStack>
 
     <VSpacer size={40} />
+
     <H2 style={{ marginBottom: 16, marginTop: 16 }}>ButtonLink</H2>
     <ComponentViewerBox name="ButtonLink · Primary variant (using Pressable API)">
       <ButtonLink
@@ -808,6 +457,7 @@ export const Buttons = () => (
         />
       </View>
     </ComponentViewerBox>
+
     <View style={styles.primaryBlock}>
       <ComponentViewerBox name="ButtonLink · Contrast variant" colorMode="dark">
         <View>
@@ -879,7 +529,9 @@ export const Buttons = () => (
         </View>
       </ComponentViewerBox>
     </View>
+
     <VSpacer size={40} />
+
     <H2 style={{ marginBottom: 16, marginTop: 16 }}>IconButton</H2>
     <ComponentViewerBox name="IconButton · Primary variant">
       <View style={IOStyles.row}>
@@ -910,6 +562,7 @@ export const Buttons = () => (
         />
       </View>
     </ComponentViewerBox>
+
     <ComponentViewerBox name="IconButton · Neutral variant">
       <View style={IOStyles.row}>
         <IconButton
@@ -942,6 +595,7 @@ export const Buttons = () => (
         />
       </View>
     </ComponentViewerBox>
+
     <View style={styles.primaryBlock}>
       <ComponentViewerBox
         name="IconButton · Contrast variant"
@@ -980,7 +634,9 @@ export const Buttons = () => (
         </View>
       </ComponentViewerBox>
     </View>
+
     <VSpacer />
+
     <View style={styles.neutralBlock}>
       <ComponentViewerBox
         name="IconButton · Neutral variant, persistent color mode"
@@ -1021,8 +677,11 @@ export const Buttons = () => (
         </View>
       </ComponentViewerBox>
     </View>
+
     <VSpacer size={40} />
+
     <H2 style={{ marginBottom: 16, marginTop: 16 }}>IconButtonSolid</H2>
+
     <ComponentViewerBox name="IconButtonSolid · Primary variant, large">
       <View style={IOStyles.row}>
         <IconButtonSolid
@@ -1045,6 +704,7 @@ export const Buttons = () => (
         />
       </View>
     </ComponentViewerBox>
+
     <View style={styles.primaryBlock}>
       <ComponentViewerBox
         name="IconButtonSolid · Contrast variant, large"
@@ -1073,7 +733,9 @@ export const Buttons = () => (
         </View>
       </ComponentViewerBox>
     </View>
+
     <VSpacer size={40} />
+
     <H2 style={{ marginBottom: 16, marginTop: 16 }}>
       IconButtonContained (Icebox)
     </H2>
@@ -1095,6 +757,7 @@ export const Buttons = () => (
         />
       </View>
     </ComponentViewerBox>
+
     <ComponentViewerBox name="IconButtonContained · Neutral variant">
       <View style={IOStyles.row}>
         <IconButtonContained
@@ -1115,6 +778,7 @@ export const Buttons = () => (
         />
       </View>
     </ComponentViewerBox>
+
     <View style={styles.primaryBlock}>
       <ComponentViewerBox
         name="IconButtonContained · Contrast variant"
@@ -1141,8 +805,11 @@ export const Buttons = () => (
         </View>
       </ComponentViewerBox>
     </View>
+
     <VSpacer size={40} />
+
     <H2 style={{ marginBottom: 16, marginTop: 16 }}>Specific buttons</H2>
+
     <ComponentViewerBox name="CalendarEventButton (using new ButtonOutline)">
       <ButtonOutline
         accessibilityHint="Tap to trigger test alert"
@@ -1160,6 +827,7 @@ export const Buttons = () => (
         onPress={onButtonPress}
       />
     </ComponentViewerBox>
+
     <ComponentViewerBox name="Login buttons">
       <ButtonSolid
         fullWidth
@@ -1187,30 +855,6 @@ const LoadingSolidButtonExample = () => {
   return (
     <>
       <ButtonSolid
-        fullWidth
-        loading={isEnabled}
-        accessibilityHint="Tap to trigger test alert"
-        label={"Primary button"}
-        onPress={() => setIsEnabled(true)}
-      />
-      <ListItemSwitch
-        label="Abilita lo stato di caricamento"
-        onSwitchValueChange={toggleSwitch}
-        value={isEnabled}
-      />
-    </>
-  );
-};
-
-const LoadingOutlineButtonExample = () => {
-  const [isEnabled, setIsEnabled] = useState(false);
-  const toggleSwitch = () => setIsEnabled(previousState => !previousState);
-
-  return (
-    <>
-      <Button
-        color="primary"
-        variant="outline"
         fullWidth
         loading={isEnabled}
         accessibilityHint="Tap to trigger test alert"
