+++ src/components/buttons/index.tsx
@@ -1,3 +1,4 @@
+export * from "./Button";
 export * from "./ButtonLink";
 export * from "./ButtonOutline";
 export * from "./ButtonSolid";
