+++ src/core/IOStyles.ts
@@ -150,6 +150,9 @@ export const IOButtonStyles = StyleSheet.create({
     elevation: 0
   },
   /* Labels */
+  label: {
+    alignSelf: "center"
+  },
   labelSizeDefault: {
     fontSize: 16
   },
